modifico el clean.sh

+++ trunk/MAPACHE/clean.sh
@@ -1,10 +1,30 @@
 #!/bin/bash
-cd build/
-$prefix_client make clean
-cd ../ 
+rm -f configPorDefecto.xml
+
+if [ -d "build/" ]
+then
+	echo "Borrando archivos de la aplicacion..."
+	cd build/
+	$prefix_client make clean
+	cd ../ 
+	rm -r -f build
+fi
+
+if [ -d "cgi-bin" ]
+then
 rm -r -f cgi-bin
+fi
+
+if [ -d "arch-temp-cgi" ]
+then
 rm -r -f arch-temp-cgi
-rm -r -f build
-rm -r -f Documentacion/Doxy
+fi
+
+if [ -d "build/" ]
+then
+	echo "Borrando documentacion..."
+	rm -r -f Documentacion/Doxy/
+fi
+
 #mkdir -p $(BASEDIR)
 #cp -r -f ../$(BASEDIR) ./
